move classes PlainText and PlainTextFormatter to org.apache.pdfbox.pdmodel.interactive
